LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCRtpParameters-codecs-expected.txt


PASS setParameters() with codec.payloadType modified should reject with InvalidModificationError (audio)
PASS setParameters() with codec.mimeType modified should reject with InvalidModificationError (audio)
PASS setParameters() with codec.clockRate modified should reject with InvalidModificationError (audio)
PASS setParameters() with codec.channels modified should reject with InvalidModificationError (audio)
PASS setParameters() with codec.sdpFmtpLine modified should reject with InvalidModificationError (audio)
PASS setParameters() with new codecs inserted should reject with InvalidModificationError (audio)
PASS setParameters() with codecs removed should reject with InvalidModificationError (audio)
PASS setParameters() with codecs reordered should reject with InvalidModificationError (audio)
FAIL setParameters() with codecs undefined should reject with InvalidModificationError (audio) promise_rejects_dom: function "function() { throw e; }" threw object "TypeError: Member RTCRtpSendParameters.codecs is required and must be an instance of sequence" that is not a DOMException InvalidModificationError: property "code" is equal to undefined, expected 13
PASS RTCRtpSender.getParameters() should not have codecs before SDP negotiation completes (audio)
PASS RTCRtpReceiver.getParameters() should not have codecs before SDP negotiation completes (audio)
FAIL RTCRtpSender.getParameters() should have codecs after negotiation (audio) assert_not_equals: got disallowed value 0
FAIL RTCRtpReceiver.getParameters() should have codecs after negotiation (audio) assert_not_equals: got disallowed value 0
FAIL RTCRtpReceiver.getParameters() codecs should match local SDP (audio, offerer) assert_not_equals: got disallowed value 0
FAIL RTCRtpReceiver.getParameters() codecs should match local SDP (audio, answerer) assert_equals: expected "minptime=10;useinbandfec=1" but got "a=fmtp:111 minptime=10;useinbandfec=1"
PASS setParameters() with codec.payloadType modified should reject with InvalidModificationError (video)
PASS setParameters() with codec.mimeType modified should reject with InvalidModificationError (video)
PASS setParameters() with codec.clockRate modified should reject with InvalidModificationError (video)
PASS setParameters() with codec.channels modified should reject with InvalidModificationError (video)
PASS setParameters() with codec.sdpFmtpLine modified should reject with InvalidModificationError (video)
PASS setParameters() with new codecs inserted should reject with InvalidModificationError (video)
PASS setParameters() with codecs removed should reject with InvalidModificationError (video)
PASS setParameters() with codecs reordered should reject with InvalidModificationError (video)
FAIL setParameters() with codecs undefined should reject with InvalidModificationError (video) promise_rejects_dom: function "function() { throw e; }" threw object "TypeError: Member RTCRtpSendParameters.codecs is required and must be an instance of sequence" that is not a DOMException InvalidModificationError: property "code" is equal to undefined, expected 13
PASS RTCRtpSender.getParameters() should not have codecs before SDP negotiation completes (video)
PASS RTCRtpReceiver.getParameters() should not have codecs before SDP negotiation completes (video)
FAIL RTCRtpSender.getParameters() should have codecs after negotiation (video) assert_not_equals: got disallowed value 0
FAIL RTCRtpReceiver.getParameters() should have codecs after negotiation (video) assert_not_equals: got disallowed value 0
FAIL RTCRtpReceiver.getParameters() codecs should match local SDP (video, offerer) assert_not_equals: got disallowed value 0
FAIL RTCRtpReceiver.getParameters() codecs should match local SDP (video, answerer) assert_equals: expected (undefined) undefined but got (number) 1
FAIL RTCRtpSender.getParameters() codecs should match remote SDP (audio, no fmtp checks, offerer) promise_test: Unhandled rejection with value: object "InvalidAccessError: Failed to set remote answer sdp: The m= section with mid='0' is invalid. RTCP-MUX is not enabled when it is required."
FAIL RTCRtpSender.getParameters() codecs should match remote SDP (audio, no fmtp checks, answerer) promise_test: Unhandled rejection with value: object "InvalidAccessError: Failed to set remote offer sdp: The m= section with mid='mid1' is invalid. RTCP-MUX is not enabled when it is required."
FAIL RTCRtpSender.getParameters() codecs should match remote SDP (audio, with fmtp checks, offerer) promise_test: Unhandled rejection with value: object "InvalidAccessError: Failed to set remote answer sdp: The m= section with mid='0' is invalid. RTCP-MUX is not enabled when it is required."
FAIL RTCRtpSender.getParameters() codecs should match remote SDP (audio, with fmtp checks, answerer) promise_test: Unhandled rejection with value: object "InvalidAccessError: Failed to set remote offer sdp: The m= section with mid='mid1' is invalid. RTCP-MUX is not enabled when it is required."
FAIL RTCRtpSender.getParameters() codecs should match remote SDP (video, minimal fmtp checks, offerer) assert_equals: expected 7 but got 9
FAIL RTCRtpSender.getParameters() codecs should match remote SDP (video, minimal fmtp checks, answerer) assert_equals: expected 7 but got 0
FAIL RTCRtpSender.getParameters() codecs should match remote SDP (video, with fmtp checks, offerer) assert_equals: expected 7 but got 9
FAIL RTCRtpSender.getParameters() codecs should match remote SDP (video, with fmtp checks, answerer) assert_equals: expected 7 but got 0
